# Vocalis-Passport/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vocalis Passport API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  confirmPassword: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export type CurrentUserRole = typeof CurrentUserRole[keyof typeof CurrentUserRole];


export const CurrentUserRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface CurrentUser {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  vocalisId: string;
  role: CurrentUserRole;
  active: boolean;
}

export interface AuthSession {
  user: CurrentUser;
}

export type StudentProfilePassportStatus = typeof StudentProfilePassportStatus[keyof typeof StudentProfilePassportStatus];


export const StudentProfilePassportStatus = {
  draft: 'draft',
  active: 'active',
} as const;

export interface StudentProfile {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  vocalisId: string;
  /** @nullable */
  profilePhotoUrl: string | null;
  level: string;
  badge: string;
  dateJoined: string;
  /** @nullable */
  dateIssued: string | null;
  passportStatus: StudentProfilePassportStatus;
  active: boolean;
}

export type PassportStatus = typeof PassportStatus[keyof typeof PassportStatus];


export const PassportStatus = {
  active: 'active',
} as const;

export interface Passport {
  id: number;
  studentId: number;
  vocalisId: string;
  fullName: string;
  /** @nullable */
  profilePhotoUrl: string | null;
  level: string;
  badge: string;
  dateJoined: string;
  dateIssued: string;
  generatedAt: string;
  status: PassportStatus;
}

export interface StudentDashboard {
  student: StudentProfile;
  passport: Passport | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  profileCompletion: number;
}

export type StudentProfileUpdateLevel = typeof StudentProfileUpdateLevel[keyof typeof StudentProfileUpdateLevel];


export const StudentProfileUpdateLevel = {
  Level_One: 'Level One',
  Level_Two: 'Level Two',
  Level_Three: 'Level Three',
  Level_Four: 'Level Four',
  Level_Five: 'Level Five',
  Level_Six: 'Level Six',
} as const;

export type StudentProfileUpdateBadge = typeof StudentProfileUpdateBadge[keyof typeof StudentProfileUpdateBadge];


export const StudentProfileUpdateBadge = {
  Explorer: 'Explorer',
  Learner: 'Learner',
  Builder: 'Builder',
  Collaborator: 'Collaborator',
  Leader: 'Leader',
  Changemaker: 'Changemaker',
  Global_Citizen: 'Global Citizen',
} as const;

export interface StudentProfileUpdate {
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  level: StudentProfileUpdateLevel;
  badge: StudentProfileUpdateBadge;
  dateJoined: string;
}

export interface PhotoUploadInput {
  photo: Blob;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlRequest;
}

export type AdminStudent = StudentProfile & {
  createdAt: string;
};

export type AdminStudentUpdateLevel = typeof AdminStudentUpdateLevel[keyof typeof AdminStudentUpdateLevel];


export const AdminStudentUpdateLevel = {
  Level_One: 'Level One',
  Level_Two: 'Level Two',
  Level_Three: 'Level Three',
  Level_Four: 'Level Four',
  Level_Five: 'Level Five',
  Level_Six: 'Level Six',
} as const;

export type AdminStudentUpdateBadge = typeof AdminStudentUpdateBadge[keyof typeof AdminStudentUpdateBadge];


export const AdminStudentUpdateBadge = {
  Explorer: 'Explorer',
  Learner: 'Learner',
  Builder: 'Builder',
  Collaborator: 'Collaborator',
  Leader: 'Leader',
  Changemaker: 'Changemaker',
  Global_Citizen: 'Global Citizen',
} as const;

export interface AdminStudentUpdate {
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /**
     * @minLength 7
     * @maxLength 30
     */
  phone: string;
  level: AdminStudentUpdateLevel;
  badge: AdminStudentUpdateBadge;
  dateJoined: string;
  /** @nullable */
  dateIssued: string | null;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Authentication required
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Insufficient permissions
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Conflict
 */
export type ConflictResponse = ErrorResponse;

/**
 * Passport data is incomplete
 */
export type UnprocessableResponse = ErrorResponse;

/**
 * Server error
 */
export type ServerErrorResponse = ErrorResponse;

export type ListAdminStudentsParams = {
/**
 * @maxLength 100
 */
search?: string;
};