# Vocalis-Passport/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vocalis Passport API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a Vocalis student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to Vocalis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current signed-in account
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request password reset instructions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/students:
    get:
      operationId: listAdminStudents
      tags: [admin]
      summary: Search registered students
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminStudent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/students/{id}:
    get:
      operationId: getAdminStudent
      tags: [admin]
      summary: View one student
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "200":
          description: Student detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateAdminStudent
      tags: [admin]
      summary: Update a student's managed fields
      parameters:
        - $ref: "#/components/parameters/StudentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminStudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/students/{id}/passport:
    post:
      operationId: generateAdminStudentPassport
      tags: [admin]
      summary: Generate a student's passport
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "200":
          description: Generated passport
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/students/{id}/passport/pdf:
    get:
      operationId: downloadAdminStudentPassportPdf
      tags: [admin]
      summary: Download a student's passport as PDF
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/students/{id}/deactivate:
    post:
      operationId: deactivateAdminStudent
      tags: [admin]
      summary: Deactivate a student account
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "200":
          description: Deactivated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    StudentId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 7
          maxLength: 30
        password:
          type: string
          minLength: 8
          maxLength: 128
        confirmPassword:
          type: string
          minLength: 8
          maxLength: 128
      required: [fullName, email, phone, password, confirmPassword]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/CurrentUser"
      required: [user]
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        vocalisId:
          type: string
        role:
          type: string
          enum: [student, admin]
        active:
          type: boolean
      required: [id, fullName, email, phone, vocalisId, role, active]
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        vocalisId:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
          format: uri
        level:
          type: string
        badge:
          type: string
        dateJoined:
          type: string
          format: date
        dateIssued:
          type: ["string", "null"]
          format: date
        passportStatus:
          type: string
          enum: [draft, active]
        active:
          type: boolean
      required:
        [id, fullName, email, phone, vocalisId, profilePhotoUrl, level, badge, dateJoined, dateIssued, passportStatus, active]
...
    Passport:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        vocalisId:
          type: string
        fullName:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
          format: uri
        level:
          type: string
        badge:
          type: string
        dateJoined:
          type: string
          format: date
        dateIssued:
          type: string
          format: date
        generatedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active]
      required:
        [id, studentId, vocalisId, fullName, profilePhotoUrl, level, badge, dateJoined, dateIssued, generatedAt, status]
    AdminStudent:
      allOf:
        - $ref: "#/components/schemas/StudentProfile"
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
          required: [createdAt]
    AdminStudentUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 7
          maxLength: 30
        level:
          type: string
          enum: [Level One, Level Two, Level Three, Level Four, Level Five, Level Six]
        badge:
          type: string
          enum: [Explorer, Learner, Builder, Collaborator, Leader, Changemaker, Global Citizen]
        dateJoined:
          type: string
          format: date
        dateIssued:
          type: ["string", "null"]
          format: date
      required: [fullName, phone, level, badge, dateJoined, dateIssued]