# Vocalis-Passport/lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vocalis Passport API
 * OpenAPI spec version: 0.1.0
 */
import type { CurrentUserRole } from './currentUserRole';

export interface CurrentUser {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  vocalisId: string;
  role: CurrentUserRole;
  active: boolean;
}