# Vocalis-Passport/lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vocalis Passport API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}