# Vocalis-Passport/lib/api-zod/src/generated/types/passport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Vocalis Passport API
 * OpenAPI spec version: 0.1.0
 */
import type { PassportStatus } from './passportStatus';

export interface Passport {
  id: number;
  studentId: number;
  vocalisId: string;
  fullName: string;
  /** @nullable */
  profilePhotoUrl: string | null;
  level: string;
  badge: string;
  dateJoined: Date;
  dateIssued: Date;
  generatedAt: Date;
  status: PassportStatus;
}